removed parameters from the column money in the Pledges model to fix error

+++ models/pledges.js
@@ -13,7 +13,7 @@ Pledges.init(
       autoIncrement: true,
     },
     money: {
-      type: DataTypes.FLOAT(1000,2),
+      type: DataTypes.FLOAT,
       allowNull: false,
       validate: {
         isNumeric: true,
